Правки модуляции
